oci: remove redundant ShouldBeStopped check for stopping containers
Fixes #8030
This commit removes the ShouldBeStopped function to address the
issue where stopping a container blocks all further stop attempts
for the same container. That function acquired the opLock
mutex to check the container's state, which could lead to contention and
potential deadlocks if other goroutines were waiting to acquire the same lock.

Additionally, the container's state is checked later in the StopContainer
function by calling the Living method, which checks if the container's
process is still running. Therefore, the ShouldBeStopped function is
redundant and can be removed.

Signed-off-by: Sohan Kunkerkar <sohank2602@gmail.com>